# backend/src/openapi/spec.yaml
openapi: 3.1.0
info:
  title: Sage Flashcards API
  description: |
    API for the Sage flashcard application with spaced repetition learning.
    Supports deck management, card creation, study sessions with FSRS algorithm,
    and AI-powered card generation.
  version: 1.0.0
  contact:
    email: support@sage.app
servers:
  - url: https://api.sage.app
    description: Production server
  - url: http://localhost:3001
    description: Development server
paths:
...
  /api/study/due:
    get:
      operationId: getDueCards
      summary: Get cards due for review
      description: Returns cards that are due for review based on FSRS scheduling
      parameters:
        - name: deckId
          in: query
          schema:
            type: string
          description: Filter by specific deck
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
          description: Maximum cards to return
      responses:
        '200':
          description: Due cards with scheduling info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DueCardsResponse'
  /api/study/review:
    post:
      operationId: submitReview
      summary: Submit a card review
      description: |
        Submit a review for a card. Rating determines the next review interval:
        - 1 (Again): Card was forgotten, will be shown again soon
        - 2 (Hard): Struggled to remember, shorter interval
        - 3 (Good): Remembered with some effort, normal interval
        - 4 (Easy): Instant recall, longer interval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cardId
                - rating
              properties:
                cardId:
                  type: string
                rating:
                  type: integer
                  enum: [1, 2, 3, 4]
                  description: "1=Again, 2=Hard, 3=Good, 4=Easy"
                reviewTimeMs:
                  type: integer
                  description: Time spent reviewing in milliseconds
      responses:
        '200':
          description: Review recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
...
components:
  schemas:
...
    Card:
      type: object
      properties:
        id:
          type: string
        front:
          type: string
        back:
          type: string
        cardType:
          type: string
          enum: [flashcard, multiple_choice]
        options:
          type: array
          items:
            type: string
        state:
          type: string
          enum: [new, learning, review, relearning]
...
    DueCardsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            cards:
              type: array
              items:
                allOf:
                  - $ref: '#/components/schemas/Card'
                  - type: object
                    properties:
                      schedule:
                        type: object
                        properties:
                          again:
                            type: object
                            properties:
                              interval:
                                type: string
                          hard:
                            type: object
                            properties:
                              interval:
                                type: string
                          good:
                            type: object
                            properties:
                              interval:
                                type: string
                          easy:
                            type: object
                            properties:
                              interval:
                                type: string
            count:
              type: integer
    ReviewResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            nextDue:
              type: string
              format: date-time
            interval:
              type: string
              description: Human-readable interval (e.g., "2 days")
...